pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "saotri-bench"
version = "0.1.0"
description = "Saotri Bench: Dynamic Coding Problems Benchmark"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
saotri-bench = "saotri_bench.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["saotri_bench*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.md"]
